+++ js/data.js
@@ -1,30 +1,5 @@
 import {getComments, getRandomArrayElement, getRandomInteger} from './util.js';
-
-const DESCRIPTIONS = [
-  'Отдыхаю',
-  'С семьёй',
-  'Красота',
-  'Работаем',
-  'Выходные'
-];
-
-const MESSAGE = [
-  'В целом всё неплохо. Но не всё.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше',
-  'Всё отлично!',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
-];
-
-const NAMES = [
-  'Маша',
-  'Петя',
-  'Даша',
-  'Никита',
-  'Артём',
-  'Витя'
-];
+import {DESCRIPTIONS} from './constants.js';
 
 function getPhotoDescription() {
   const PHOTO_DESCRIPTIONS = [];
@@ -42,4 +17,4 @@ function getPhotoDescription() {
   return PHOTO_DESCRIPTIONS;
 }
 
-export {getPhotoDescription, MESSAGE, NAMES};
+export {getPhotoDescription};
