+++ js/main.js
@@ -1,2 +1,2 @@
-import './util.js';
-import './data.js';
+import {getPhotoDescription} from './data.js';
+getPhotoDescription();
