+++ js/util.js
@@ -1,4 +1,5 @@
-import {NAMES, MESSAGE} from './data.js';
+import {NAMES, MESSAGE} from './constants.js';
+
 const getRandomInteger = (a, b) => {
   const lower = Math.ceil(Math.min(a, b));
   const upper = Math.floor(Math.max(a, b));
